Created a separate file for module tests and fixed some linter issues.
