+++ app/scripts/main.js
@@ -1,5 +1,3 @@
 'use strict';
 
 require('../modules/modals/scripts.js');
-
-import Pestle from './pestle/main.js';
