+++ app/scripts/pestle/main.js
@@ -1,3 +1,5 @@
+'use strict';
+
 import Module from './lib/module';
 
 class Pestle {
